+++ src/components/Battery/Battery.tsx
@@ -101,7 +101,7 @@ const SideText = ({
                                 ? `${Math.min(
                                       100,
                                       Math.max(latestAdcSample.soc ?? 0, 0)
-                                  ).toFixed(1)} %`
+                                  ).toFixed(1)}%`
                                 : 'N/A %'}
                         </h2>
                     </DocumentationTooltip>