+++ src/components/Profiling/Dialog/CompleteDialog.tsx
@@ -327,7 +327,7 @@ export default ({ isVisible }: { isVisible: boolean }) => {
         <GenericDialog
             title={`Battery Profiling ${
                 profile.name.length > 0 ? `- ${profile.name}` : ''
-            } @ ${profile.temperatures[index]} °C`}
+            } @ ${profile.temperatures[index]}°C`}
             isVisible={isVisible}
             showSpinner={generatingBatteryModel}
             closeOnEsc={false}